Moved dc-ipa-01 post install script into own post install directory.
The dc-ipa-01 post install script now documents a complete freeipa server. But it isn't fully automated yet. It needs to have the main ipa-server-install process automated through an answer file. Once it is truly automated, the commands can be appended to the kickstart file. Ultimately the kick start file should have all the required commands, and the virt-install command creates the vm with predefined settings, and uses the ks file to do the rest.