Added libvirt server, and qemu conf since modifications were required.
After installing the vdsm package on the KVM physical host, it appears the package locked down libvirt. I had to make modifications to the configuration files to allow myself to install VMs the way I previously had. The uncommented lines in the added files are what was required to get things working properly again.